reporting/requirements.txt
boto3==1.34.129
botocore==1.34.129
jmespath==1.0.1
lxml==5.2.2
numpy==2.0.0
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
